pyproject.toml
[tool.poetry]
name = "extractor"
version = "0.1.0"
description = "No-code data extractor"
authors = ["Maria <maria@dataqa.ai>", "Stuart <stuart@dataqa.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0" # needed for streamlit
pandas = "2.1.3"
python-dotenv = "1.0.0"
pydantic = "2.4.2"
openai = "0.28.0" # https://stackoverflow.com/questions/77540822)
instructor = "0.2.11" # last version to depend on openai < 1.0.0
streamlit = "1.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
